Fix SHA1 implementation and enable unit test
The defined uint32 type must be exactly 32 bit wide.
Otherwise the SHA1::finalBin() method will break.

+++ rutil/Sha1.hxx
@@ -43,7 +43,7 @@ public:
     static std::string from_file(const std::string &filename);
 
 private:
-    typedef unsigned long int uint32;   /* just needs to be at least 32bit */
+    typedef unsigned int uint32;        /* needs to be exactly 32bit */
     typedef unsigned long long uint64;  /* just needs to be at least 64bit */
 
     static const unsigned int DIGEST_INTS = 5;  /* number of 32bit integers per SHA1 digest */
